renaming irc to bancho
