+++ BanchoBot/commands/example.ts
